refactor: 레이어드 아키텍처로 구조 개선
