+++ docs/요구사항-2.md
@@ -26,3 +26,20 @@
 - 출력
   - 검사 기록을 출력한다
     - 검사 결과를 색깔로 변환하여 출력한다.
+
+---
+
+### 구조
+- domain
+  - Letter 
+    - Word
+  - Result
+  - Round
+  - GameRecords
+    - GameRecord
+  - lines => Word
+- view
+  - GameView
+- util
+  - FileReader
+- GameMachine
