+++ src/main/java/wordle/GameMachine.java
@@ -1,5 +1,7 @@
 package wordle;
 
+import wordle.util.FileReader;
+import wordle.domain.Result;
 import wordle.view.GameView;
 
 import java.io.BufferedReader;
