+++ src/main/java/wordle/domain/GameRecord.java
@@ -1,4 +1,4 @@
-package wordle;
+package wordle.domain;
 
 public class GameRecord {
 }
