+++ src/main/java/wordle/domain/GameRecords.java
@@ -1,4 +1,4 @@
-package wordle;
+package wordle.domain;
 
 public class GameRecords {
 }
