+++ src/main/java/wordle/domain/Result.java
@@ -1,6 +1,6 @@
-package wordle;
+package wordle.domain;
 
-enum Result {
+public enum Result {
     정답, // 위치와 문자가 맞다 (초록색)
     문자만_같음,// 위치는 틀리고 문자는 맞다 (노란색)
     틀림, // 둘다 틀렸다 (회색)
