+++ src/main/java/wordle/domain/Round.java
@@ -1,4 +1,4 @@
-package wordle;
+package wordle.domain;
 
 public class Round {
