+++ src/main/java/wordle/util/FileReader.java
@@ -1,4 +1,4 @@
-package wordle;
+package wordle.util;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
