+++ src/test/java/wordle/FileReaderTest.java
@@ -2,6 +2,7 @@ package wordle;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wordle.util.FileReader;
 
 import java.util.List;
