+++ src/test/java/wordle/GameMachineTest.java
@@ -1,6 +1,7 @@
 package wordle;
 
 import org.junit.jupiter.api.Test;
+import wordle.domain.Result;
 
 import java.util.List;
