+++ src/test/java/wordle/WordsTest.java
@@ -2,6 +2,7 @@ package wordle;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wordle.domain.Word;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
